feat(bundler): Bundlers have common implementation methods
ss.bundler provides common functionality
