+++ lib/client/bundler/default.js
@@ -34,15 +34,22 @@ function mkdir(dir) {
 }
 
 
-module.exports = function bundler(ss,client,options){
+module.exports = function(ss,client,options){
   var bundler = {
     define: define,
     load: load,
+    asset: {
+      js: assetJS,
+      worker: assetWorker,
+      css: assetCSS,
+      html: assetHTML
+    },
+    pack: {
+      js: packJS,
+      css: packCSS
+    },
     ensureAssetFolder: ensureAssetFolder
   };
-
-  bundler.pack = require('./pack')(ss,client,bundler,options);
-  bundler.asset = require('./asset')(ss,options);
 
   return bundler;
 
@@ -76,25 +83,41 @@ module.exports = function bundler(ss,client,options){
 	}
 
   function load() {
-    var containerDir = path.join(ss.root, options.dirs.assets);
-    var clientDir = path.join(containerDir, client.name);
+    this.description = ss.bundler.descriptionFor(ss,client,options);
+  }
 
-    this.description = {
+  function assetJS(path, opts, cb) {
+    return ss.bundler.loadFile(ss, options.dirs.code, path, 'js', opts, function(output) {
+      output = ss.bundler.wrapCode(output, path, opts.pathPrefix, options);
+      if (opts.compress && path.indexOf('.min') === -1) {
+        output = ss.bundler.minifyJSFile(output, path);
+      }
+      return cb(output);
+    });
+  }
+  function assetWorker(path, opts, cb) {
+    return ss.bundler.loadFile(ss, options.dirs.workers, path, 'js', opts, function(output) {
+      if (opts.compress) {
+        output = ss.bundler.minifyJSFile(output, path);
+      }
+      return cb(output);
+    });
+  }
 
-      //TODO perhaps mixin the abs versions by SS
-      paths: {
-        html: path.join(clientDir, client.id + '.html'),
-        js: path.join(clientDir, client.id + '.js'),
-        css: path.join(clientDir, client.id + '.css')
-      },
-      relPaths: {
-        html: path.join(options.dirs.assets, client.name, client.id + '.html'),
-        js: path.join(options.dirs.assets, client.name, client.id + '.js'),
-        css: path.join(options.dirs.assets, client.name, client.id + '.css')
-      },
-      dir: clientDir,
-      containerDir: containerDir
-    };
+  function assetCSS(path, opts, cb) {
+    return ss.bundler.loadFile(ss, options.dirs.css, path, 'css', opts, cb);
+  }
+
+  function assetHTML(path, opts, cb) {
+    return ss.bundler.loadFile(ss, options.dirs.views, path, 'html', opts, cb);
+  }
+
+  function packJS(postProcess) {
+    ss.bundler.packAssetSet('js', options.dirs.code, client, bundler, postProcess);
+  }
+
+  function packCSS(postProcess) {
+    ss.bundler.packAssetSet('css', options.dirs.css, client, bundler, postProcess);
   }
 
   //TODO move to index.js
