+++ lib/client/bundler/index.js
@@ -2,10 +2,15 @@
 'use strict';
 
 var fs = require('fs'),
+    path = require('path'),
     log = require('../../utils/log'),
     cleanCSS = require('clean-css'),
     system = require('../system'),
-    view = require('../view');
+    view = require('../view'),
+    wrap = require('../wrap'),
+    uglifyjs = require('uglify-js'),
+    jsp = uglifyjs.parser,
+    pro = uglifyjs.uglify;
 
 /**
  * Bundler by client name
@@ -75,7 +80,7 @@ exports.pack = function pack(ss, client, options) {
     var minified, original;
     original = files.join('\n');
     minified = cleanCSS.process(original);
-    log.info(('  Minified CSS from ' + (formatKb(original.length)) + ' to ' + (formatKb(minified.length))).grey);
+    log.info(('  Minified CSS from ' + (exports.formatKb(original.length)) + ' to ' + (exports.formatKb(minified.length))).grey);
     return minified;
   });
 
@@ -100,8 +105,153 @@ exports.pack = function pack(ss, client, options) {
   });
 };
 
-// PRIVATE
+// API for implementing bundlers
 
-function formatKb(size) {
-  return '' + (Math.round((size / 1024) * 1000) / 1000) + ' KB';
+exports.loadFile = function loadFile(ss, dir, fileName, type, options, cb) {
+  dir = path.join(ss.root, dir);
+  var p = path.join(dir, fileName);
+  var extension = path.extname(p);
+  extension = extension && extension.substring(1); // argh!
+  var formatter = ss.client.formatters[extension];
+  if (p.substr(0, dir.length) !== dir) {
+    throw new Error('Invalid path. Request for ' + p + ' must not live outside ' + dir);
+  }
+  if (!formatter) {
+    throw new Error('Unsupported file extension \'.' + extension + '\' when we were expecting some type of ' + (type.toUpperCase()) + ' file. Please provide a formatter for ' + (p.substring(ss.root.length)) + ' or move it to /client/static');
+  }
+  if (formatter.assetType !== type) {
+    throw new Error('Unable to render \'' + fileName + '\' as this appears to be a ' + (formatter.assetType.toUpperCase()) + ' file. Expecting some type of ' + (type.toUpperCase()) + ' file in ' + (dir.substr(ss.root.length)) + ' instead');
+  }
+  return formatter.compile(p.replace(/\\/g, '/'), options, cb);
+};
+
+exports.minifyJSFile = function minifyJSFile(originalCode, fileName) {
+  var ast = jsp.parse(originalCode);
+  ast = pro.ast_mangle(ast);
+  ast = pro.ast_squeeze(ast);
+  var minifiedCode = pro.gen_code(ast);
+  log.info(('  Minified ' + fileName + ' from ' + (exports.formatKb(originalCode.length)) + ' to ' + (exports.formatKb(minifiedCode.length))).grey);
+  return minifiedCode;
+};
+
+exports.descriptionFor = function(ss,client,options) {
+  var containerDir = path.join(ss.root, options.dirs.assets);
+  var clientDir = path.join(containerDir, client.name);
+
+  return {
+
+    //TODO perhaps mixin the abs versions by SS
+    paths: {
+      html: path.join(clientDir, client.id + '.html'),
+      js: path.join(clientDir, client.id + '.js'),
+      css: path.join(clientDir, client.id + '.css')
+    },
+    relPaths: {
+      html: path.join(options.dirs.assets, client.name, client.id + '.html'),
+      js: path.join(options.dirs.assets, client.name, client.id + '.js'),
+      css: path.join(options.dirs.assets, client.name, client.id + '.css')
+    },
+    dir: clientDir,
+    containerDir: containerDir
+  };
 }
+
+exports.packAssetSet = function packAssetSet(assetType, dir, client, bundler, postProcess) {
+  var filePaths,
+    prefix,
+    paths = client.paths[assetType];
+
+  function writeFile(fileContents) {
+    var fileName = bundler.description.paths[assetType];
+    fs.writeFileSync(fileName, postProcess(fileContents));
+    return log.info('✓'.green, 'Packed', filePaths.length, 'files into', bundler.description.relPaths[assetType]);
+  }
+
+  function processFiles(fileContents, i) {
+    var file, path, _ref;
+    if (!fileContents) {
+      fileContents = [];
+    }
+    if (!i) {
+      i = 0;
+    }
+    _ref = filePaths[i], path = _ref.path, file = _ref.file;
+    return bundler.asset[assetType](file, {
+      pathPrefix: path,
+      compress: true
+    }, function(output) {
+      fileContents.push(output);
+      if (filePaths[++i]) {
+        return processFiles(fileContents, i);
+      } else {
+        return writeFile(fileContents);
+      }
+    });
+  }
+
+  // Expand any dirs into real files
+  if (paths && paths.length > 0) {
+    filePaths = [];
+    prefix = path.join(ss.root, dir);
+    paths.forEach(function(path) {
+      return magicPath.files(prefix, path).forEach(function(file) {
+        return filePaths.push({
+          path: path,
+          file: file
+        });
+      });
+    });
+    return processFiles();
+  }
+}
+
+
+exports.formatKb = function formatKb(size) {
+  return '' + (Math.round((size / 1024) * 1000) / 1000) + ' KB';
+};
+
+// Before client-side code is sent to the browser any file which is NOT a library (e.g. /client/code/libs)
+// is wrapped in a module wrapper (to keep vars local and allow you to require() one file in another).
+// The 'system' directory is a special case - any module placed in this dir will not have a leading slash
+exports.wrapCode = function wrapCode(code, path, pathPrefix, options) {
+  var modPath, sp, i;
+  var pathAry = path.split('/');
+
+  // Don't touch the code if it's in a 'libs' directory
+  if (pathAry.indexOf('libs') >= 0) {
+    return code;
+  }
+
+  if (pathAry.indexOf('entry.js') === -1 && options && options.browserifyExcludePaths) {
+    for(i in options.browserifyExcludePaths) {
+      if (options.browserifyExcludePaths.hasOwnProperty(i)) {
+        if ( path.split( options.browserifyExcludePaths[i] )[0] === '' ) {
+          return code;
+        }
+      }
+    }
+  }
+
+  // Don't add a leading slash if this is a 'system' module
+  if (pathAry.indexOf('system') >= 0) {
+    modPath = pathAry[pathAry.length - 1];
+    return wrap.module(modPath, code);
+  } else {
+
+    // Otherwise treat as a regular module
+    modPath = options.globalModules? pathAry.join("/") : pathAry.slice(1).join('/');
+
+    // Work out namespace for module
+    if (pathPrefix) {
+
+      // Ignore any filenames in the path
+      if (pathPrefix.indexOf('.') > 0) {
+        sp = pathPrefix.split('/');
+        sp.pop();
+        pathPrefix = sp.join('/');
+      }
+      modPath = path.substr(pathPrefix.length + 1);
+    }
+    return wrap.module('/' + modPath, code);
+  }
+};
