+++ lib/client/index.js
@@ -39,6 +39,9 @@ var clients = {};
 
 module.exports = function(ss, router) {
 
+  // make bundler methods available for default and other implementations
+  ss.bundler = bundler;
+
   // Require sub modules
   var templateEngine = require('./template_engine')(ss),
       formatters = require('./formatters')(ss),
